+++ client/src/pages/index.js
@@ -8,4 +8,3 @@ export * from './Payment';
 export * from './SignIn';
 export * from './SignUp';
 export * from './ViewTutorial';
-export * from './WhoWeAre';
